+++ R/utils.R
@@ -111,7 +111,7 @@ new_total_colors <- c(
 #'
 #' @returns countries shapefile
 #'
-load_countries_data <- function(destpath = system.file("./countries_data", package = "Covid19")){
+load_countries_data_map <- function(destpath = system.file("./countries_data", package = "Covid19")){
   # Resource https://www.naturalearthdata.com/downloads/50m-cultural-vectors/50m-admin-0-countries-2/
   url <- "https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip"
   zip_path <- file.path(destpath,"ne_50m_admin_0_countries.zip")
@@ -158,12 +158,12 @@ align_country_names_pop <- function(data){
 
       "Antigua and Barbuda" = "Antigua and Barb.",
       "Bosnia and Herzegovina" = "Bosnia and Herz.",
-      "Cape Verde" = "Cabo Verde",
+      #"Cape Verde" = "Cabo Verde",
       "Cote d'Ivoire" = "C\\u00f4te d'Ivoire",
       "Czech Republic" = "Czechia",
       "Dominican Republic" = "Dominican Rep.",
       "Eswatini" = "eSwatini",
-      "French Guiana" = "Guyana",
+      #"French Guiana" = "Guyana",
       "North Macedonia" = "Macedonia",
       "UK" = "United Kingdom",
       "USA" = "United States of America",
@@ -182,12 +182,12 @@ align_country_names_pop_reverse <- function(data){
 
       "Antigua and Barb." = "Antigua and Barbuda",
       "Bosnia and Herz." = "Bosnia and Herzegovina",
-      "Cabo Verde" = "Cape Verde",
+      #"Cabo Verde" = "Cape Verde",
       "C\\u00f4te d'Ivoire" = "Cote d'Ivoire",
       "Czechia" = "Czech Republic",
       "Dominican Rep." = "Dominican Republic",
       "eSwatini" = "Eswatini",
-      "Guyana" = "French Guiana",
+      #"Guyana" = "French Guiana",
       "Macedonia" = "North Macedonia",
       "United Kingdom" = "UK",
       "United States of America" = "USA",
@@ -207,4 +207,40 @@ clean_plotly_leg <- function(.plotly_x, .extract_str) {
     .plotly_x$name <- stringr::str_extract(.plotly_x$name, .extract_str)
   }
   .plotly_x
+}
+#' Aggregates data to continent or subcontinent (group)
+#' @param data data.frame aggregated data per Country.Region
+#' @param group character continent or subcontinent
+#' @param time character date or contagion_day
+#' @param popdata data with population info
+#' @param allstatuses character vector of statuses to base the recomputation from: confirmed recovered deaths active
+#'
+#' @note growth and mortality variables must be recomputed after the aggregation
+#' the return dataset renames the group column into Country.region to allow the following graphs to work
+#'
+#' @return data.frame aggregated at group level
+#'
+#' @importFrom rlang sym
+#' @export
+aggr_to_cont = function(data, group, time, popdata, allstatuses) {
+
+  popdata_cont = popdata %>% filter(!is.na(!!rlang::sym(group))) %>%
+    group_by(.dots = group) %>%
+    summarize(population = sum(population, rm.na = T))
+
+  continent_data =    data %>%
+    select(Country.Region, population, contagion_day, date, !!group, date, !!allstatuses) %>%
+    mutate(population = as.numeric(population)) %>%
+    group_by(.dots = c(time,group)) %>%
+    #group_by(time, continent) %>%
+    summarise_at(c(allstatuses), sum, na.rm = TRUE) %>%
+    add_growth_death_rate(group, time) %>%
+    left_join(popdata_cont[,c(group, "population")], by = group) %>%
+    mutate(mortality_rate_1M_pop = round(10^6*deaths/population, digits = 3),
+           prevalence_rate_1M_pop = round(10^6*confirmed/population, digits = 3),
+           new_prevalence_rate_1M_pop = round(10^6*new_confirmed/population, digits = 3)) %>%
+    rename(Country.Region = !!group) %>%
+    get_timeseries_by_contagion_day_data()  %>%
+    arrange(desc(date))
+  continent_data
 }