+++ man-roxygen/ex-continents.R
@@ -1,20 +1,20 @@
 if (interactive()) {
   library(shiny)
-  library(Covid19)
   library(dplyr)
+  library(Covid19)
   library(tidyr)
-  library(ggplot2)
 
   long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
   ui <- fluidPage(
-    Covid19:::mod_bar_plot_day_contagion_ui("bar_plot_day_contagion")
+    tabPanel("Continents",
+             tabsetPanel(
+               tabPanel("Summary",
+                        id = "tab_global",
+    Covid19:::mod_continent_comparison_ui("continent_comparison")
+               )))
   )
   server <- function(input, output) {
 
-    n <- 1000 #  min number of cases for a country to be considered. Default 1000
-    w <- 7 # number of days of outbreak. Default 7
-
-    # Data ----
     orig_data <- reactive({
       get_timeseries_full_data() %>%
         get_timeseries_by_contagion_day_data()
@@ -34,22 +34,18 @@ if (interactive()) {
                new_prevalence_rate_1M_pop = round(10^6*new_confirmed/population, digits = 3))
       orig_data_aggregate
     })
-
+    n = 1000; w = 7
     data_filtered <- reactive({
       orig_data_aggregate() %>%
         Covid19:::rescale_df_contagion(n = n, w = w)
     })
 
-    country_data <- reactive({
-      data_filtered() %>%
-        filter(Country.Region %in% "Switzerland") %>%
-        filter(contagion_day > 0) %>%
-        arrange(desc(date))
-    })
-
-
-
-    callModule(Covid19:::mod_bar_plot_day_contagion_server,"bar_plot_day_contagion", country_data = country_data)
+    # countries <- reactive({
+    #   data_filtered() %>%
+    #     select(Country.Region) %>%
+    #     distinct()
+    # })
+    callModule(Covid19:::mod_continent_comparison_server, "continent_comparison", orig_data_aggregate = orig_data_aggregate, data_filtered = data_filtered, n = n, w = w, pop_data)
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }