+++ man-roxygen/ex-mod_stackedbarplot_status.R
@@ -14,12 +14,14 @@ if (interactive()) {
   )
   server <- function(input, output, session) {
 
+    pop_data = get_pop_data()
+
     orig_data_aggregate = reactive({get_timeseries_full_data() %>%
       get_timeseries_by_contagion_day_data() %>%
       aggregate_province_timeseries_data() %>%
       arrange(Country.Region) %>%
         align_country_names_pop() %>%
-        get_pop_data() %>% # compute additional variables
+        merge_pop_data(pop_data) %>% # compute additional variables
         align_country_names_pop_reverse() %>%
         mutate(mortality_rate_1M_pop = round(10^6*deaths/population, digits = 3),
                prevalence_rate_1M_pop = round(10^6*confirmed/population, digits = 3),