+++ NEWS.md
@@ -1,6 +1,6 @@
 ### Covid19 1.1.2.9000 (development branch)
 - correct plot title in country page (#91)
-
+- Added continent split pages (#91)
 
 ### Covid19 1.1.7 (2020-05-28)